+++ app/src/api-connectors/manufacturers.js
