+++ app/src/api-connectors/statistics.js
